Break games page tests into multiple files to help with test isolation
